VCHI/MMAL: ported VCHI and MMAL from linux kernel
VCHI - VideoCore interface, that allows speaking to RTOS's running on
a VideoCore GPU. This interface in essense is a message transport to deliver
messages to or from the core.
VCHIQ - Q in the end stands for queue, way of organizing shared memory in a
message queue.

Built on it are several services:
1. "MEMS" - memory service for allocating and managing memory buffers between
   ARM and GPU,
2. "MMAL" - Multi-Media-Abstraction-Layer, that implements functionality to
   access cameras and displays in generic way.

VCHIQ setup procedure is the following:
Allocate sufficient chunk of contignuous memory, which will be split into
'slots' later.
In the first part of memory chunk insert a VCHIQ descriptor, also called
"slot zero".
Initialize values of "slot zero" slot to describe memory layout of the whole
chunk, but also VCHIQ protocol version and some status flags
Pass address to the whole memory chunk to VideoCore via MBOX.
VideoCore will receive address to memory chunk and parse "slot zero" at the
start of this chunk.
After that VCHI is considered connected and running.
Communication is established by splitting whole memory chunk (excluding first
part for "slot zero") into 2 different direction messages - to GPU / from GPU
, remote / local. Both sides would have own descriptors, pointers and indices
to the actual data slots in the memory and other fields indicating the address
of a new inserted message.
Also both sides of communication ARM and GPU make use of simple synchronization
primitives, made up from structures in "slot zero" and ARM doorbell mechanism,
As soon as one places the new message into the memory, it can set flag on
the correspoding event structure and signal BELL0 or BELL1 to generate
interrupt on the other size.
Upon receiving and interrupt the accepring side would wake up and check for the
reason for being interrupted - what exact flag was triggered (or fired),
and then lookup the latest message from the shared memory.