vchiq: started implementation of block IO scheduling
Goal is to write H264 compressed video frames to sdcard.

H264 compression is provided by Raspberry PI VideoCore GPIO firmware called
VCOS (VideoCore Operating System) by its big subsystem MMAL (Multi Media
Abstraction Layer). This is implemented as a various components, each having
a set of inputs and outputs, that can be connected into graphs.
In our case component "ril.camera" provides a video output port, that is
constantly capturing video data and is ready to output this to the consumer,
this camera port output is connected to encoder component "ril.video_encode",
which consumes raw video frames from camera video port, compresses it according
to H264 routines and provides to our OS in series of BUFFER_TO_HOST messages.

In order to retreive compressed data from encoder component, we have to
provide buffers to encoder. "ril.video_encode" requires 1 64K buffer. We send
this buffer with BUFFER_FROM_HOST message. When encoder is ready to give us
some data - it returns it with BUFFER_TO_HOST message. From this moment encoder
has not buffers and waits for us to process the buffer and give the buffer
back to encoder. This process is ongoing. Encoder compresses video frames,
stores compressed info into buffer, sends to us, then it has no buffer, waits
for the buffer to use from us, as soon as it gets the buffer it writes next
compressed data to it, gives that back to us with BUFFER_TO_HOST message and so
on.
We in turn provide encoder with a buffer, wait until encoder returns this
buffer filled with data back to us, use this buffer in some way like write out
to file or to other buffer and then return the buffer back to encoder with
BUFFER_FROM_HOST message, doing this on and on in an ongoing loop.

Couple of important details here:
- Encoder asks one buffer of size 64k. But we can provide bigger one just
  in order to reduce fragmentation, so 512k is the buffer size now.
- Buffer, sent back to us, rarely occupy whole buffer capacity, usually single
  messages are much smaller than buffer max size, like 28 bytes for some H264
  internal header or 3567 bytes of compressed data, etc.
- Writing all those buffer data in sequential order into a contignous data
  block or a file will result in a perfectly valid video file, if saved on
  disk - it will be possible to play that with any video player.

The challenge here is that as soon as camera video port is enabled and capture
is requested once, encoder will continuosly send us new data in buffers, and
if we fail to give back ther buffers to encoder it will drop some of the
data, we will never write it and basicly loose important chunks, resulting in
glitches in the final video file. This is especially important in our case
because we can not just save compressed video information to RAM, we are
storing all of that to an sdcard, which might be slow to write to.

Previos implementation had blocking calls to write data to sdcard, whenever
enough bytes have been received from encoder. So encoder sends the data, we
collect it to some bigger buffer which size is in sectors. As soon this buffer
is filled this data gets written to sdcard in blocking mode. Because sd card
is slow, some of the ongoing encoder data gets lost.

In this change the idea is to let encoder do it's work as smooth as possible.
Encoder only works with 1 buffer at a time. So we provide a buffer with
BUFFER_FROM_HOST. Encoder works with that, then sends the buffer back to us.
At this point we give it another free preallocated buffer, while leaving
the buffer we just received for ourselves for a little big longer. This way
Encoder is unblocked to continue, while we have bought some time to process
encoder's data.

Next in the solution is that we maintain 2 buffers, both having a big size
in multiples of 512 bytes (sector sizE). While 1 buffer is full with encoder
data, we request the whole buffer to be flushed to sdcard, so for the time
of flushing this buffer becomes unavailable for further copying, for that time
we use the second buffer filling it with data, coming from encoder, while
sdcard is written with the contents of other buffer in parallel.

To support that, I have created blockdev_scheduler. For now its just a fancy
word, but actually its pretty simple - it runs IO requests to sd card block
device in its own thread and runs callbacks when work is done.
The callback just indicates that job is done and buffer is free for new
mem-copying from ecnoder buffers.