bcm2835_emmc: added stream write interface
This is to experiment with non stop writes to sd card. bcm2835 emmc controller
supports writing continuosly, For this CMD23 SET_BLOCK_COUNT should set
maximum possible number of blocks, right after that CMD25 WRITE_MULTIPLE_BLOCKS
should be sent, with same number of blocks set in BLKSIZECOUNT register, after
that, it is possible to write to its DATA register contignuosly. DATA register
seems to have 16k fifo, so after 16k/4 writes are performed the contents of a
16k fifo buffer is flushed to the SD card.