Camera stack: video out+splitter+resizer+encoder work somehow
Updated component graph for the MMAL camera stack:
- Component 'ril.camera' captures video frames and exports video stream in
  OPAQUE format via VIDEO_OUT port in 1280x1024 resolution.
- Component 'ril.video_splitter' splits OPAQUE video stream to two equal
  streams and exports then via ports OUT0 and OUT1 in OPAQUE format
- Component 'ril.isp' is here to reduce frame resolution to 320x240 for SPI
  display and outputs via OUT port as raw RGB24 320x240 frames
- Port 'ril.camera.VIDEO_OUT' is connected to port 'splitter.IN'
- Port 'splitter.OUT0' goes to port 'encoder.IN'
- Port 'splitter.OUT1' goes to port 'ril.isp.IN'
- Port 'ril.isp.OUT' outputs raw pixel stream in RGB24 to SPI display output
- Port 'encoder.OUT' outputs H264 compressed stream to SD card controller for
  contignous write

                                        ___________            ____________
     _________       __________        |  (h264)   |          | (h264.mov) |
    |         |     |          |------>|  ENCODER  |--------->|  SD CARD   |
    |  VIDEO  |---->| SPLITTER |       |___________|          |____________|
    |_________|     |__________|-\      ___________            ____________
                                  \    |  ril.isp  |          |  (RGB24)   |
                                   --->|  RESIZER  |--------->| SPI DISPLAY|
                                       |___________|          |____________|

'ril.isp' is used as resizer instead of 'ril.resize' in a role of resizer,
for two reasons:
- 'ril.resizer' does not support RGB24 - which is 3 bytes per pixel,
red, green and blue, and by default suggests using RGBA format (4 bytes, fourth
is alpha), which is not very useful for writing to SPI display, which also
works with 24 bit RGB pixels. We can just feed results of 'ril.isp' resizing
in RGB24 straight to DMA SPI transfer and nothing else is required. In case of
RGBA we would have to manually get rid of A channel by processing the whole
320x240x3 buffer.
- 'ril.resizer' is implemented as software solution, highly optimized by using
  vector instructions, while 'ril.isp' is a more hardware accelerated solution
  and should give better performance, this I haven't yet had a chance to check

Also all components are using zero copying.

Possible ways to improve the current graph would be to enable 'ril.camera'
PREVIEW port and connect it to 'ril.isp', removing splitter completely