Memory: dma memory now differs from device memory
Previously MMU was setup similarly for both device memory and dma memory.
The difference is rather big though. Device memory means each IO to separate
addresses is a completely unique operation, that can not be combined, merged,
bursted, omitted, etc. On the other hand dma memory is basicly a memory
without cacheing, so that DMA peripheral could write to this area to and from
other memory or device locations and it doesn't need to be cache invalidated.
ARMv8 architecture describes several memory attributes that can be assigned
to memory region, that give useful hints to the underlying memory bus in terms
of how to optimize memory access. GRE - Gather, Reorder, Early-Write-Ack
So while for device memory we should go for the most restrictive memory attrs
nGnRnE - where all of the above is disallowed, for the dma memory it is ok
to have them and can possibly give some performance boost